#487: Reordered release and github_client files for calypso
